+++ extension/test/unit/lib/logger.test.js
@@ -16,12 +16,12 @@ describe('Logger', () => {
   it('should log the status code, duration, request and response', () => {
     const logSpy = sinon.spy(log, 'debug')
 
-    logRequest.log(httpCodeSuccess, {}, {}, {})
+    logRequest.log({}, {statusCode: httpCodeSuccess, elapsedTime: 123})
 
     sinon.assert.calledWith(logSpy, sinon.match.has('duration'))
-    sinon.assert.calledWith(logSpy, sinon.match.has('message'))
-    sinon.assert.calledWith(logSpy, sinon.match.has('request'))
-    sinon.assert.calledWith(logSpy, sinon.match.has('response'))
+    sinon.assert.calledWith(logSpy, sinon.match.has('shopifyRequest'))
+    sinon.assert.calledWith(logSpy, sinon.match.hasNested('shopifyRequest.response'))
+    sinon.assert.calledWith(logSpy, sinon.match.hasNested('shopifyRequest.request'))
     sinon.assert.calledWith(logSpy, sinon.match({statusCode: httpCodeSuccess}))
   })
 })