+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Shopify request log format changed to be compatible with `shopify-user`'s log format
 
 ## [1.2.6] - 2018-09-28
 ### Fixed